Update script to work with Bind 9.10 V3 stats
Bind 9.10 has new XML stats Version 3. This update detects this and uses
the new XML layout for stats collection.

It also adds memory usage and cache data for Version 3 and Memory for
Version 2.

The Zabbix Template for Zabbix 3.x is updates with the new data fields
for memory and cache, also adding graphs for those.

The discovory rules is disabled, as it seems to create too much noise
and would need some regex detect update.